+++ enlighten/timing/BatchCollection.py
@@ -86,7 +86,7 @@ class BatchCollection:
     
     @par Laser Control
     
-    There are three use-cases:
+    There are four use-cases:
     
     - MANUAL: means the user is manually controlling the laser (if there
       is one), so don't worry about it
@@ -94,11 +94,13 @@ class BatchCollection:
       off at the end (measurement_count)
     - SPECTRUM: turn the laser on at the beginning of a measurement (each
       count), off after.
+    - AUTO-RAMAN: use the standard ENLIGHTEN Auto-Raman feature to generate
+      one or more an optimized measurements at each batch cycle
     
-    The three modes are exposed in the configuration file as:
+    The four modes are exposed in the configuration file as:
     
     @verbatim
-      laser_mode = manual | batch | spectrum
+      laser_mode = manual | batch | spectrum | auto-raman
     @endverbatim
     
     If multiple spectrometers are connected, the laser commands (like acquisition
@@ -113,21 +115,9 @@ class BatchCollection:
     
     It's clearly useful to be able to take a fresh dark at the start of each
     batch, and that has been implemented.
-    
-    However, another customer requested the ability to take a dark before each
-    measurement within a batch (presumably with laser mode "Spectrum"), and that's
-    a bit more tricky.  Right now "spectrum" laser mode is controlled inside
-    Wasatch.PY, so for each measurement we'd have to disable the driver's laser mode,
-    TakeOne dark, set the dark, then re-enable laser mode, and take the measurement.
-    That's not much better than having manual control of the laser from this
-    process, which is what we were trying to avoid, as it doesn't really allow
-    fine-grained control over warmup time.
-    
-    For now, I'm following in the architectural path of pushing somewhat more
-    functionality down into Wasatch.PY, and letting the driver provide BOTH the
-    dark and the laser control around the Reading.  We won't be performing dark
-    subtraction within the driver, but adding a .dark attribute to the Reading
-    which ENLIGHTEN can register and store upon receipt.
+
+    If you need a fresh dark before each measurement (not just each batch), 
+    then Auto-Raman mode is recommended.
     """
     def __init__(self, ctl):
         self.ctl = ctl
@@ -141,6 +131,7 @@ class BatchCollection:
         self.rb_laser_manual                = cfu.radioButton_BatchCollection_laser_manual
         self.rb_laser_spectrum              = cfu.radioButton_BatchCollection_laser_spectrum
         self.rb_laser_batch                 = cfu.radioButton_BatchCollection_laser_batch
+        self.rb_laser_auto_raman            = cfu.radioButton_BatchCollection_laser_auto_raman
         self.spinbox_measurement_count      = cfu.spinBox_BatchCollection_measurement_count
         self.spinbox_measurement_period_ms  = cfu.spinBox_BatchCollection_measurement_period_ms
         self.spinbox_batch_count            = cfu.spinBox_BatchCollection_batch_count
@@ -189,6 +180,7 @@ class BatchCollection:
         self.rb_laser_manual               .toggled      .connect(self.update_from_widgets)
         self.rb_laser_spectrum             .toggled      .connect(self.update_from_widgets)
         self.rb_laser_batch                .toggled      .connect(self.update_from_widgets)
+        self.rb_laser_auto_raman           .toggled      .connect(self.update_from_widgets)
         self.spinbox_measurement_count     .valueChanged .connect(self.update_from_widgets)
         self.spinbox_measurement_period_ms .valueChanged .connect(self.update_from_widgets)
         self.spinbox_batch_count           .valueChanged .connect(self.update_from_widgets)
@@ -250,6 +242,8 @@ class BatchCollection:
                 self.rb_laser_batch.setChecked(True)
             elif self.laser_mode == "spectrum":
                 self.rb_laser_spectrum.setChecked(True)
+            elif self.laser_mode == "auto_raman":
+                self.rb_laser_auto_raman.setChecked(True)
 
     def update_from_widgets(self):
 
@@ -268,6 +262,8 @@ class BatchCollection:
             self.laser_mode = "spectrum"
         elif self.rb_laser_batch.isChecked():
             self.laser_mode = "batch"
+        elif self.rb_laser_auto_raman.isChecked():
+            self.laser_mode = "auto_raman"
         else: 
             self.laser_mode = "manual"
 
@@ -312,6 +308,10 @@ class BatchCollection:
 
             if self.laser_mode == "manual":
                 s += "<p>The laser will not be automatically turned on or off during the collection (<b>Laser Mode Manual</b>).</p>"
+            elif self.laser_mode == "auto_raman":
+                s += """<p>Every measurement within the batch will use the standard ENLIGHTEN 'Auto-Raman' feature, 
+                           meaning acquisition parameters will be optimized according to the current Auto-Raman 
+                           configuration settings to generate an ideal averaged, dark-corrected Raman measurement.</p>"""
             else:
                 s += f"<p>The laser will be automatically turned on at the beginning of each {self.laser_mode} (<b>Laser Mode</b>). "
                 if self.dark_before_batch:
@@ -339,7 +339,6 @@ class BatchCollection:
             return False
 
         self.ctl.vcr_controls.register_observer("stop", self.vcr_stop)
-        #self.ctl.vcr_controls.register_observer("save", self.save_complete)
         self.collection_start_time = datetime.datetime.now()
         if self.spinbox_collection_timeout.value() != 0:
             self.collection_timeout = self.collection_start_time + datetime.timedelta(seconds=self.spinbox_collection_timeout.value())
@@ -352,7 +351,11 @@ class BatchCollection:
             avg = self.ctl.scan_averaging.get_scans_to_average() # #432
 
             # initialize driver-level laser auto_triggering and dark collection
-            if self.laser_mode == "spectrum":
+            if self.laser_mode == "auto_raman":
+                aar = self.ctl.auto_raman.generate_auto_raman_request()
+                self.take_one_template = TakeOneRequest(auto_raman_request=aar)
+                self.ctl.auto_raman.force_on(True)
+            elif self.laser_mode == "spectrum":
                 self.take_one_template = TakeOneRequest(scans_to_average=avg, take_dark=self.dark_before_batch, enable_laser_before=True, disable_laser_after=True, laser_warmup_ms=self.laser_warmup_ms)
             else:
                 self.take_one_template = TakeOneRequest(scans_to_average=avg)
@@ -508,6 +511,7 @@ class BatchCollection:
         if self.laser_mode != "manual":
             log.debug(f"BatchCollection.stop: disabling all lasers (laser_mode {self.laser_mode})")
             self.ctl.laser_control.set_laser_enable(False, all_=True)
+        self.ctl.auto_raman.force_on(False)
 
         self.ctl.vcr_controls.unregister_observer("save", self.save_complete)
         self.ctl.vcr_controls.unregister_observer("stop", self.vcr_stop)